fix big numbers

+++ printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: rtruvelo <rtruvelo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/16 14:45:14 by rtruvelo          #+#    #+#             */
-/*   Updated: 2023/12/22 11:27:28 by rtruvelo         ###   ########.fr       */
+/*   Updated: 2024/01/25 14:00:17 by rtruvelo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 // 	struct s_list	*next;
 // }	t_list;
 
-int		ft_atoi(const char	*str);
+long		ft_atoi(const char	*str);
 void	*ft_bzero(void *s, size_t n);
 size_t	ft_strlen(const char *str);
 char	*ft_itoa(int n);
